+++ core/src/main/java/org/apache/calcite/runtime/SqlFunctions.java
@@ -71,7 +71,6 @@ import java.util.Objects;
 import java.util.Set;
 import java.util.TimeZone;
 import java.util.concurrent.atomic.AtomicLong;
-import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 import javax.annotation.Nonnull;
 
@@ -548,78 +547,6 @@ public class SqlFunctions {
 
     int flags = caseSensitive ? 0 : Pattern.CASE_INSENSITIVE;
     return Pattern.compile(regex, flags).matcher(s).find();
-  }
-
-  /** SQL {@code REGEXP_REPLACE} function. */
-  public static String regexpReplace(String s, String regex, String replacement) {
-    return regexpReplace(s, regex, replacement, 1, 0, null);
-  }
-
-  /** SQL {@code REGEXP_REPLACE} function. */
-  public static String regexpReplace(String s, String regex, String replacement,
-                                     Integer pos) {
-    return regexpReplace(s, regex, replacement, pos, 0, null);
-  }
-
-  /** SQL {@code REGEXP_REPLACE} function. */
-  public static String regexpReplace(String s, String regex, String replacement,
-                                     Integer pos, Integer occurrence) {
-    return regexpReplace(s, regex, replacement, pos, occurrence, null);
-  }
-
-  /** SQL {@code REGEXP_REPLACE} function. */
-  public static String regexpReplace(String s, String regex, String replacement,
-                                     Integer pos, Integer occurrence, String matchType) {
-    if (pos < 1 || pos > s.length()) {
-      throw RESOURCE.invalidInputForRegexpReplace(pos.toString()).ex();
-    }
-
-    int count = 0;
-    StringBuffer sb = new StringBuffer();
-    String input = s;
-    if (pos != 1) {
-      sb.append(s.substring(0, pos - 1));
-      input = s.substring(pos - 1);
-    }
-
-    int flags = makeRegexpFlags(matchType);
-    Matcher matcher = Pattern.compile(regex, flags).matcher(input);
-    while (matcher.find()) {
-      if (occurrence == 0) {
-        matcher.appendReplacement(sb, replacement);
-      } else if (++count == occurrence) {
-        matcher.appendReplacement(sb, replacement);
-        break;
-      }
-    }
-    matcher.appendTail(sb);
-
-    return sb.toString();
-  }
-
-  private static int makeRegexpFlags(String stringFlags) {
-    int flags = 0;
-    if (stringFlags != null) {
-      for (int i = 0; i < stringFlags.length(); ++i) {
-        switch (stringFlags.charAt(i)) {
-          case 'i':
-            flags |= Pattern.CASE_INSENSITIVE;
-            break;
-          case 'c':
-            flags &= ~Pattern.CASE_INSENSITIVE;
-            break;
-          case 'n':
-            flags |= Pattern.DOTALL;
-            break;
-          case 'm':
-            flags |= Pattern.MULTILINE;
-            break;
-          default:
-            throw RESOURCE.invalidInputForRegexpReplace(stringFlags).ex();
-        }
-      }
-    }
-    return flags;
   }
   // =
