+++ core/src/main/resources/org/apache/calcite/runtime/CalciteResource.properties
@@ -290,4 +290,5 @@ InvalidInputForJsonLength=Not a valid input for JSON_LENGTH: ''{0}''
 InvalidInputForJsonKeys=Not a valid input for JSON_KEYS: ''{0}''
 InvalidInputForJsonRemove=Invalid input for JSON_REMOVE: document: ''{0}'', jsonpath expressions: ''{1}''
 InvalidInputForJsonStorageSize=Not a valid input for JSON_STORAGE_SIZE: ''{0}''
+InvalidInputForRegexpReplace=Not a valid input for REGEXP_REPLACE: ''{0}''
 # End CalciteResource.properties
